Integer -> Long 수정

+++ src/main/java/com/example/review/Repository/ReviewRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ReviewRepository extends JpaRepository<Review, Long> {
-    Page<Review> findByHospitalId(Integer id, Pageable pageable);
+    Page<Review> findByHospitalId(Long id, Pageable pageable);
 }
