+++ src/main/java/com/example/review/controller/HospitalRestController.java
@@ -29,13 +29,13 @@ public class HospitalRestController {
 
 
     @GetMapping("/{id}")
-    public ResponseEntity<HospitalResponse> get(@PathVariable Integer id) { // ResponseEntity도 DTO타입
+    public ResponseEntity<HospitalResponse> get(@PathVariable Long id) { // ResponseEntity도 DTO타입
         HospitalResponse hospitalResponse = hospitalService.getHospital(id); // DTO
         return ResponseEntity.ok().body(hospitalResponse); // Return은 DTO로
     }
 
     @GetMapping("/{id}/reviews")
-    public ResponseEntity<Page<Review>> getReviews(@PathVariable Integer hospitalId) {
+    public ResponseEntity<Page<Review>> getReviews(@PathVariable Long hospitalId) {
         Pageable pageable = Pageable.ofSize(20);
         return ResponseEntity.ok().body(reviewService.get(hospitalId, pageable));
     }
